+++ gitclone/gitclone.go
@@ -1,0 +1,14 @@
+package gitclone
+
+import (
+	"github.com/WJQSERVER-STUDIO/go-utils/logger"
+)
+
+var (
+	logw       = logger.Logw
+	logDump    = logger.LogDump
+	logDebug   = logger.LogDebug
+	logInfo    = logger.LogInfo
+	logWarning = logger.LogWarning
+	logError   = logger.LogError
+)